Add copy of margins to rows and image
These will need to be changed. The idea here is to push them to relative units.
Also may need to indroduce things like grid and flex later on in development.
for right now the goal is to replicate the site to an extent, then modify
the behavior to better fit the needs of this project. Things like loading
a JS function for each tab, or having a dropdown for breakfast and dinner
menus.