Place dinner text image in container and make flex object
This was done to center the text image between the two menus. It also required
me to go into a photo-editior(GIMP in this case) and crop/edit the image down
to size. This will likely have to be done with all text based images, depending
on where we want them.