Create zipped list for drinks and descriptions
This concludes the structure for the drinks menu. There should be descriptions
and comments added for future code legibility before moving onto breakfast
and dinner menues. The current code may also be refined before publication.
It currently isn't using any factory function or IIFE's and therefore
a lot of the data isn't private.