Change apps FF to IIFE
This is being done in order to potentially streamline a lot of redundanct calls
and to improve code structure and cohesiveness. While each subsection of the
menu has it's own unique elements, many of the elements are standard and
can be implemented within a single series of functions, rather than
an entire page of separate functions all handling the same things.