Returns array from linked webpage
The code is largely from @theotrain
https://github.com/theotrain/load-google-sheet-as-csv-with-js

This code will be modified as needed. I may decide to use a sql
database later, or keep the XLSX and parse it manually. For now this works
with the code I've generated, and should be ready to go in implementation of
menu functions.