Make fetchData fetch multiple sheets and return them
Each fetched sheet is pushed to a data array in main, which is then assigned
multiple variables in menu.js. This adds a certain level of insecurity
for the actual loading of the site. It should be a practice to download
and overwrite the csv spread sheet upon each load-for the admin. That
way if the site fails to communitcate with google sheets there is a
backup locally to load from