Link to all pages and add scroll events
This commit not only adds all the proper loading of pages, but also
ensures that when clicked, the page is returned to the top. This
change was made because clicking the contact link scrolls to the
bottom of the page, and when loading a different page would
remain in place. The reloading of the page still needs to
be fixed such that it sticks to the current page.