Fixes contact from duplicating
This commit checks to see if the footer is already there, and
does append to the body if it is. Also in this commit, the
main div container was converted to a main container, and
the footer is appended to the body, not the main. Given
how the code functions, the contact page should never be
replaced, making that check a bit redundant. However, it is
left in place for now as code around anchoring pages is
ongoing.