Fix page reload to home
This commit fixes the defaulting to home upon the user refreshing their
browser. Instead, now the url is obtained, and the last part of the
url link is split into an array, and the second part is assigned to the
variable currentPage. If the user refreshes the page, when the IIFE...executes
the variable is passed to the loadPage function. If currentPage is undefined,
we default it to 'home'.